backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
groq
PyPDF2
gTTS
python-dotenv
opencv-python
mediapipe
numpy
chromadb
sentence-transformers
reportlab
python-jose[cryptography]
bcrypt
aiosqlite
sqlalchemy